+++ src/components/OverviewList/BudgetsOverview/BudgetsOverview.css
@@ -1,0 +1,119 @@
+.BudgetOverviewContainer {
+  width: 428px;
+  height: 410px;
+  background: var(--white);
+  border-radius: 12px;
+  padding: 32px;
+  display: flex;
+  flex-direction: column;
+  overflow: hidden;
+}
+
+.BudgetHeader {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 10px;
+}
+
+.BudgetHeader h2 {
+  font-size: 18px;
+  color: var(--grey-900);
+  font-weight: bold;
+}
+
+.SeeDetails {
+  background: none;
+  border: none;
+  color: var(--grey-900);
+  font-weight: bold;
+  cursor: pointer;
+  font-size: 14px;
+}
+
+.SeeDetails:hover {
+  text-decoration: underline;
+}
+
+.BudgetContent {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  flex: 1;
+  overflow: hidden;
+}
+
+.BudgetChart {
+  width: 250px;
+  height: 250px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  position: relative;
+}
+
+.BudgetChart canvas {
+  max-width: 100%;
+  max-height: 100%;
+}
+
+.BudgetChartCenter {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  text-align: center;
+}
+
+.BudgetUsed {
+  font-size: 24px;
+  font-weight: bold;
+  color: var(--grey-900);
+}
+
+.BudgetLimit {
+  font-size: 12px;
+  color: var(--grey-500);
+}
+
+.BudgetList {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  max-height: 300px;
+  overflow-y: auto;
+  flex-wrap: wrap;
+  align-items: flex-start;
+  justify-content: center;
+}
+
+.BudgetItem {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  min-width: 100px;
+}
+
+.BudgetColor {
+  width: 4px;
+  height: 43px;
+  border-radius: 4px;
+}
+
+.BudgetInfo {
+  display: flex;
+  flex-direction: column;
+}
+
+.BudgetInfo p {
+  font-size: 14px;
+  color: var(--grey-900);
+}
+
+.BudgetInfo strong {
+  font-size: 16px;
+  color: var(--grey-900);
+}
