+++ src/pages/Overview/Overview.css
@@ -4,10 +4,10 @@
   gap: 20px;
   align-items: center;
   justify-content: flex-start;
-  padding: 32px; /* 전체 페이지 패딩 */
+  padding: 32px;
   width: 100%;
-  height: 100vh; /* 전체 화면을 차지하도록 설정 */
-  box-sizing: border-box; /* 패딩이 width, height를 초과하지 않도록 설정 */
+  height: 100vh;
+  box-sizing: border-box;
 }
 
 .OverviewHeader {
@@ -18,128 +18,4 @@
 .OverviewHeader h1 {
   font-weight: bold;
   color: var(--grey-900);
-}
-
-.BudgetOverviewContainer {
-  width: 428px;
-  height: 410px;
-  background: var(--white);
-  border-radius: 12px;
-  padding: 32px;
-  display: flex;
-  flex-direction: column;
-  overflow: hidden;
-}
-
-.BudgetHeader {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 10px;
-}
-
-.BudgetHeader h2 {
-  font-size: 18px;
-  color: var(--grey-900);
-  font-weight: bold;
-}
-
-.SeeDetails {
-  background: none;
-  border: none;
-  color: var(--grey-900);
-  font-weight: bold;
-  cursor: pointer;
-  font-size: 14px;
-}
-
-.SeeDetails:hover {
-  text-decoration: underline;
-}
-
-.BudgetContent {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  flex: 1;
-  overflow: hidden;
-}
-
-.BudgetChart {
-  width: 300px;
-  height: 300px;
-}
-
-.BudgetList {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  max-height: 300px;
-  overflow-y: auto;
-  flex-wrap: wrap;
-  align-items: flex-start;
-  justify-content: center;
-}
-
-.BudgetItem {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-  min-width: 100px;
-}
-
-.BudgetColor {
-  width: 4px;
-  height: 43px;
-  border-radius: 4px;
-}
-
-.BudgetInfo {
-  display: flex;
-  flex-direction: column;
-}
-
-.BudgetInfo p {
-  font-size: 14px;
-  color: var(--grey-900);
-}
-
-.BudgetInfo strong {
-  font-size: 16px;
-  color: var(--grey-900);
-}
-.BudgetChart {
-  width: 250px;
-  height: 250px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  position: relative;
-}
-
-.BudgetChart canvas {
-  max-width: 100%;
-  max-height: 100%;
-}
-
-.BudgetChartCenter {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  text-align: center;
-}
-
-.BudgetUsed {
-  font-size: 24px;
-  font-weight: bold;
-  color: var(--grey-900);
-}
-
-.BudgetLimit {
-  font-size: 12px;
-  color: var(--grey-500);
 }
